backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.28
pydantic>=2
python-dotenv>=1.0
SQLAlchemy>=2.0
pandas>=2.0
numpy>=1.26
scikit-learn>=1.3
statsmodels>=0.14
pytest>=7
pytest-cov>=4
httpx>=0.24
